Fix allocation error discovered by gcc 7.1.
The integer value `-1` was being returned from functions that get the
workspace's number of send or receive partners in certain cases. This value was
then being passed as a `size_t` into an allocation function. Unfortunately, `-1`
is a very large `size_t` and would generate issues. I adjusted this so that `0`
is returned in these corner cases rather than `MAX_SIZE_T`.

+++ include/pgf_fe2_server_rebalance.h
@@ -68,7 +68,7 @@ int* pgf_FE2_server_rebalance_keep_buf(const pgf_FE2_server_rebalance *t);
 /**
  * Get the number of jobs to send.
  */
-int pgf_FE2_server_rebalance_n_send(const pgf_FE2_server_rebalance *t);
+size_t pgf_FE2_server_rebalance_n_send(const pgf_FE2_server_rebalance *t);
 
 /**
  * Get pointer to beginning of buffer containing job ids to send.
@@ -84,7 +84,7 @@ int* pgf_FE2_server_rebalance_send_dest(const pgf_FE2_server_rebalance *t);
 /**
  * Get the number of jobs to receive.
  */
-int pgf_FE2_server_rebalance_n_recv(const pgf_FE2_server_rebalance *t);
+size_t pgf_FE2_server_rebalance_n_recv(const pgf_FE2_server_rebalance *t);
 
 /**
  * Get pointer to beginning of buffer containing job ids to recv.
